refactor(api): extract shared OpenAPI codegen module; tests import source of truth (review #2)
Code-review theme (security + testing + maintainability): the byte-faithful
serializer was copy-pasted across the openapi-inject-* injectors, and the
contract tests re-derived the injector's helpers / scraped its source with
duplicate regexes — a tautology that catches a dropped injection but not a wrong
value, plus latent drift (required.mjs's eq had already diverged).

Extract scripts/lib/openapi-codegen.mjs (pure node, zero npm deps — safe in the
make-generate context): the serializer (sortRec/goEscape/serialize/eq),
normalizeKey, the fail-closed gateway/entitlement/premium source-of-truth
parsers, and PUBLIC_FORBIDDEN_GATES. Rewire openapi-inject-{security,examples,
servers}.mjs to import it, and rewire the security + examples contract tests to
import the SAME parsers/gates/normalizeKey instead of re-scraping the injector —
so the test can no longer drift from the injector via a divergent private regex.

Pure refactor: make generate is byte-identical (0 diff across docs/api +
src/generated), all 3 injectors --check idempotent, contract tests pass.
